requirements.txt
python-dotenv~=1.0.0
Flask~=2.2.3
twilio~=7.17.0
google-cloud-dialogflow~=2.21.0